tomorrow i will finish
https://www.youtube.com/watch?v=biOSgUjGB_o

This video helped me the most, now I have to go through everything again and try to explain all the steps
Also I need to study how to plot somthing with more than 2 PCA, in this case the number of PCAs is 8 and it is too much to plot a graph.

I want to try with the PCA(0.95), with 95% of information and variance

And I need to find the best model to use in this case.