Explaining the steps
I have started to explain all the steps in my project. Today I found out that maybe it is not a good idea to substitute all the missing values for the mean or the median because it could be impact considerably. Now the challenge is to find scientific base to support the drop to some columns that have a significant number of missing values. After lunch I will experiment with these.