Normalization
Previously I had used the standard scaler, but in today’s class I have learnt that for a sparse distribution the l2 is better. Another thing that I have found is that I was doing the histogram in the right way. For tomorrow I have to finish all the details and review.