Changing
Today I changed my imputation method from median to the knn. I normalised the data using the l2 method. The test with dropping rows and the one dropping columns are ready, I just need to reference them in the main file. Today I discovered that i wasn’t referencing in the Harvard style.

For tomorrow: reference the tests, change the referencings for Harvard, Justify the ML, and write the conclusion.